Crate rusoto_appstream

Amazon AppStream 2.0
This is the Amazon AppStream 2.0 API Reference. It provides descriptions and syntax for each of the actions and data types in AppStream 2.0. AppStream 2.0 is a fully managed application streaming service. You centrally manage your desktop applications on AppStream 2.0 and securely deliver them to any computer. AppStream 2.0 manages the AWS resources required to host and run your applications, scales automatically, and provides access to your users on demand.

To learn more about AppStream 2.0, see the following resources:

Amazon AppStream 2.0 product page

Amazon AppStream 2.0 documentation

If you're using the service, you're probably looking for AppStreamClient and AppStream.

Structs

AppStreamClient
A client for the Amazon AppStream API.

Application
Describes an application in the application catalog.

ApplicationSettings
The persistent application settings for users of a stack.

ApplicationSettingsResponse
Describes the persistent application settings for users of a stack.

AssociateFleetRequest

AssociateFleetResult

BatchAssociateUserStackRequest

BatchAssociateUserStackResult

BatchDisassociateUserStackRequest

BatchDisassociateUserStackResult

ComputeCapacity
Describes the capacity for a fleet.

ComputeCapacityStatus
Describes the capacity status for a fleet.

CopyImageRequest

CopyImageResponse

CreateDirectoryConfigRequest

CreateDirectoryConfigResult

CreateFleetRequest

CreateFleetResult

CreateImageBuilderRequest

CreateImageBuilderResult

CreateImageBuilderStreamingURLRequest

CreateImageBuilderStreamingURLResult

CreateStackRequest

CreateStackResult

CreateStreamingURLRequest

CreateStreamingURLResult

CreateUsageReportSubscriptionRequest

CreateUsageReportSubscriptionResult

CreateUserRequest

CreateUserResult

DeleteDirectoryConfigRequest

DeleteDirectoryConfigResult

DeleteFleetRequest

DeleteFleetResult

DeleteImageBuilderRequest

DeleteImageBuilderResult

DeleteImagePermissionsRequest

DeleteImagePermissionsResult

DeleteImageRequest

DeleteImageResult

DeleteStackRequest

DeleteStackResult

DeleteUsageReportSubscriptionRequest

DeleteUsageReportSubscriptionResult

DeleteUserRequest

DeleteUserResult

DescribeDirectoryConfigsRequest

DescribeDirectoryConfigsResult

DescribeFleetsRequest

DescribeFleetsResult

DescribeImageBuildersRequest

DescribeImageBuildersResult

DescribeImagePermissionsRequest

DescribeImagePermissionsResult

DescribeImagesRequest

DescribeImagesResult

DescribeSessionsRequest

DescribeSessionsResult

DescribeStacksRequest

DescribeStacksResult

DescribeUsageReportSubscriptionsRequest

DescribeUsageReportSubscriptionsResult

DescribeUserStackAssociationsRequest

DescribeUserStackAssociationsResult

DescribeUsersRequest

DescribeUsersResult

DirectoryConfig
Describes the configuration information required to join fleets and image builders to Microsoft Active Directory domains.

DisableUserRequest

DisableUserResult

DisassociateFleetRequest

DisassociateFleetResult

DomainJoinInfo
Describes the configuration information required to join fleets and image builders to Microsoft Active Directory domains.

EnableUserRequest

EnableUserResult

ExpireSessionRequest

ExpireSessionResult

Fleet
Describes the parameters for a fleet.

FleetError
Describes a fleet error.

Image
Describes an image.

ImageBuilder
Describes a virtual machine that is used to create an image.

ImageBuilderStateChangeReason
Describes the reason why the last image builder state change occurred.

ImagePermissions
Describes the permissions for an image.

ImageStateChangeReason
Describes the reason why the last image state change occurred.

LastReportGenerationExecutionError
Describes the error that is returned when a usage report can't be generated.

ListAssociatedFleetsRequest

ListAssociatedFleetsResult

ListAssociatedStacksRequest

ListAssociatedStacksResult

ListTagsForResourceRequest

ListTagsForResourceResponse

NetworkAccessConfiguration
Describes the network details of the fleet instance for the streaming session.

ResourceError
Describes a resource error.

ServiceAccountCredentials
Describes the credentials for the service account used by the fleet or image builder to connect to the directory.

Session
Describes a streaming session.

SharedImagePermissions
Describes the permissions that are available to the specified AWS account for a shared image.

Stack
Describes a stack.

StackError
Describes a stack error.

StartFleetRequest

StartFleetResult

StartImageBuilderRequest

StartImageBuilderResult

StopFleetRequest

StopFleetResult

StopImageBuilderRequest

StopImageBuilderResult

StorageConnector
Describes a connector that enables persistent storage for users.

TagResourceRequest

TagResourceResponse

UntagResourceRequest

UntagResourceResponse

UpdateDirectoryConfigRequest

UpdateDirectoryConfigResult

UpdateFleetRequest

UpdateFleetResult

UpdateImagePermissionsRequest

UpdateImagePermissionsResult

UpdateStackRequest

UpdateStackResult

UsageReportSubscription
Describes information about the usage report subscription.

User
Describes a user in the user pool.

UserSetting
Describes an action and whether the action is enabled or disabled for users during their streaming sessions.

UserStackAssociation
Describes a user in the user pool and the associated stack.

UserStackAssociationError
Describes the error that is returned when a user can’t be associated with or disassociated from a stack.

VpcConfig
Describes VPC configuration information for fleets and image builders.

Enums

AssociateFleetError
Errors returned by AssociateFleet

BatchAssociateUserStackError
Errors returned by BatchAssociateUserStack

BatchDisassociateUserStackError
Errors returned by BatchDisassociateUserStack

CopyImageError
Errors returned by CopyImage

CreateDirectoryConfigError
Errors returned by CreateDirectoryConfig

CreateFleetError
Errors returned by CreateFleet

CreateImageBuilderError
Errors returned by CreateImageBuilder

CreateImageBuilderStreamingURLError
Errors returned by CreateImageBuilderStreamingURL

CreateStackError
Errors returned by CreateStack

CreateStreamingURLError
Errors returned by CreateStreamingURL

CreateUsageReportSubscriptionError
Errors returned by CreateUsageReportSubscription

CreateUserError
Errors returned by CreateUser

DeleteDirectoryConfigError
Errors returned by DeleteDirectoryConfig

DeleteFleetError
Errors returned by DeleteFleet

DeleteImageBuilderError
Errors returned by DeleteImageBuilder

DeleteImageError
Errors returned by DeleteImage

DeleteImagePermissionsError
Errors returned by DeleteImagePermissions

DeleteStackError
Errors returned by DeleteStack

DeleteUsageReportSubscriptionError
Errors returned by DeleteUsageReportSubscription

DeleteUserError
Errors returned by DeleteUser

DescribeDirectoryConfigsError
Errors returned by DescribeDirectoryConfigs

DescribeFleetsError
Errors returned by DescribeFleets

DescribeImageBuildersError
Errors returned by DescribeImageBuilders

DescribeImagePermissionsError
Errors returned by DescribeImagePermissions

DescribeImagesError
Errors returned by DescribeImages

DescribeSessionsError
Errors returned by DescribeSessions

DescribeStacksError
Errors returned by DescribeStacks

DescribeUsageReportSubscriptionsError
Errors returned by DescribeUsageReportSubscriptions

DescribeUserStackAssociationsError
Errors returned by DescribeUserStackAssociations

DescribeUsersError
Errors returned by DescribeUsers

DisableUserError
Errors returned by DisableUser

DisassociateFleetError
Errors returned by DisassociateFleet

EnableUserError
Errors returned by EnableUser

ExpireSessionError
Errors returned by ExpireSession

ListAssociatedFleetsError
Errors returned by ListAssociatedFleets

ListAssociatedStacksError
Errors returned by ListAssociatedStacks

ListTagsForResourceError
Errors returned by ListTagsForResource

StartFleetError
Errors returned by StartFleet

StartImageBuilderError
Errors returned by StartImageBuilder

StopFleetError
Errors returned by StopFleet

StopImageBuilderError
Errors returned by StopImageBuilder

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

UpdateDirectoryConfigError
Errors returned by UpdateDirectoryConfig

UpdateFleetError
Errors returned by UpdateFleet

UpdateImagePermissionsError
Errors returned by UpdateImagePermissions

UpdateStackError
Errors returned by UpdateStack

Traits

AppStream
Trait representing the capabilities of the Amazon AppStream API. Amazon AppStream clients implement this trait.