Struct rusoto_appstream::ImageBuilder

pub struct ImageBuilder {
    pub appstream_agent_version: Option<String>,
    pub arn: Option<String>,
    pub created_time: Option<f64>,
    pub description: Option<String>,
    pub display_name: Option<String>,
    pub domain_join_info: Option<DomainJoinInfo>,
    pub enable_default_internet_access: Option<bool>,
    pub image_arn: Option<String>,
    pub image_builder_errors: Option<Vec<ResourceError>>,
    pub instance_type: Option<String>,
    pub name: String,
    pub platform: Option<String>,
    pub state: Option<String>,
    pub state_change_reason: Option<ImageBuilderStateChangeReason>,
    pub vpc_config: Option<VpcConfig>,
}

Describes a streaming instance used for editing an image. New images are created from a snapshot through an image builder.

Fields
appstream_agent_version: Option<String>

The version of the AppStream 2.0 agent that is currently being used by this image builder.

arn: Option<String>

The ARN for the image builder.

created_time: Option<f64>

The time stamp when the image builder was created.

description: Option<String>

The description for display.

display_name: Option<String>

The image builder name for display.

domain_join_info: Option<DomainJoinInfo>

The information needed to join a Microsoft Active Directory domain.

enable_default_internet_access: Option<bool>

Enables or disables default internet access for the image builder.

image_arn: Option<String>

The ARN of the image from which this builder was created.

image_builder_errors: Option<Vec<ResourceError>>

The image builder errors.

instance_type: Option<String>

The instance type for the image builder.

name: String

The name of the image builder.

platform: Option<String>

The operating system platform of the image builder.

state: Option<String>

The state of the image builder.

state_change_reason: Option<ImageBuilderStateChangeReason>

The reason why the last state change occurred.

vpc_config: Option<VpcConfig>

The VPC configuration of the image builder.