Trait rusoto_appstream::AppStream

pub trait AppStream {
    fn associate_fleet(
        &self, 
        input: &AssociateFleetRequest
    ) -> Result<AssociateFleetResult, AssociateFleetError>;
    fn create_fleet(
        &self, 
        input: &CreateFleetRequest
    ) -> Result<CreateFleetResult, CreateFleetError>;
    fn create_stack(
        &self, 
        input: &CreateStackRequest
    ) -> Result<CreateStackResult, CreateStackError>;
    fn create_streaming_url(
        &self, 
        input: &CreateStreamingURLRequest
    ) -> Result<CreateStreamingURLResult, CreateStreamingURLError>;
    fn delete_fleet(
        &self, 
        input: &DeleteFleetRequest
    ) -> Result<DeleteFleetResult, DeleteFleetError>;
    fn delete_stack(
        &self, 
        input: &DeleteStackRequest
    ) -> Result<DeleteStackResult, DeleteStackError>;
    fn describe_fleets(
        &self, 
        input: &DescribeFleetsRequest
    ) -> Result<DescribeFleetsResult, DescribeFleetsError>;
    fn describe_images(
        &self, 
        input: &DescribeImagesRequest
    ) -> Result<DescribeImagesResult, DescribeImagesError>;
    fn describe_sessions(
        &self, 
        input: &DescribeSessionsRequest
    ) -> Result<DescribeSessionsResult, DescribeSessionsError>;
    fn describe_stacks(
        &self, 
        input: &DescribeStacksRequest
    ) -> Result<DescribeStacksResult, DescribeStacksError>;
    fn disassociate_fleet(
        &self, 
        input: &DisassociateFleetRequest
    ) -> Result<DisassociateFleetResult, DisassociateFleetError>;
    fn expire_session(
        &self, 
        input: &ExpireSessionRequest
    ) -> Result<ExpireSessionResult, ExpireSessionError>;
    fn list_associated_fleets(
        &self, 
        input: &ListAssociatedFleetsRequest
    ) -> Result<ListAssociatedFleetsResult, ListAssociatedFleetsError>;
    fn list_associated_stacks(
        &self, 
        input: &ListAssociatedStacksRequest
    ) -> Result<ListAssociatedStacksResult, ListAssociatedStacksError>;
    fn start_fleet(
        &self, 
        input: &StartFleetRequest
    ) -> Result<StartFleetResult, StartFleetError>;
    fn stop_fleet(
        &self, 
        input: &StopFleetRequest
    ) -> Result<StopFleetResult, StopFleetError>;
    fn update_fleet(
        &self, 
        input: &UpdateFleetRequest
    ) -> Result<UpdateFleetResult, UpdateFleetError>;
    fn update_stack(
        &self, 
        input: &UpdateStackRequest
    ) -> Result<UpdateStackResult, UpdateStackError>;
}

Trait representing the capabilities of the Amazon AppStream API. Amazon AppStream clients implement this trait.

Required Methods

fn associate_fleet(
&self,
input: &AssociateFleetRequest
) -> Result<AssociateFleetResult, AssociateFleetError>

Associate a fleet to a stack.

fn create_fleet(
&self,
input: &CreateFleetRequest
) -> Result<CreateFleetResult, CreateFleetError>

Creates a new fleet.

fn create_stack(
&self,
input: &CreateStackRequest
) -> Result<CreateStackResult, CreateStackError>

Create a new stack.

fn create_streaming_url(
&self,
input: &CreateStreamingURLRequest
) -> Result<CreateStreamingURLResult, CreateStreamingURLError>

Creates a URL to start an AppStream 2.0 streaming session for a user. By default, the URL is valid only for 1 minute from the time that it is generated.

fn delete_fleet(
&self,
input: &DeleteFleetRequest
) -> Result<DeleteFleetResult, DeleteFleetError>

Deletes a fleet.

fn delete_stack(
&self,
input: &DeleteStackRequest
) -> Result<DeleteStackResult, DeleteStackError>

Deletes the stack. After this operation completes, the environment can no longer be activated, and any reservations made for the stack are released.

fn describe_fleets(
&self,
input: &DescribeFleetsRequest
) -> Result<DescribeFleetsResult, DescribeFleetsError>

If fleet names are provided, this operation describes the specified fleets; otherwise, all the fleets in the account are described.

fn describe_images(
&self,
input: &DescribeImagesRequest
) -> Result<DescribeImagesResult, DescribeImagesError>

Describes the images. If a list of names is not provided, all images in your account are returned. This operation does not return a paginated result.

fn describe_sessions(
&self,
input: &DescribeSessionsRequest
) -> Result<DescribeSessionsResult, DescribeSessionsError>

Describes the streaming sessions for a stack and a fleet. If a user ID is provided, this operation returns streaming sessions for only that user. Pass this value for the nextToken parameter in a subsequent call to this operation to retrieve the next set of items. If an authentication type is not provided, the operation defaults to users authenticated using a streaming URL.

fn describe_stacks(
&self,
input: &DescribeStacksRequest
) -> Result<DescribeStacksResult, DescribeStacksError>

If stack names are not provided, this operation describes the specified stacks; otherwise, all stacks in the account are described. Pass the nextToken value in a subsequent call to this operation to retrieve the next set of items.

fn disassociate_fleet(
&self,
input: &DisassociateFleetRequest
) -> Result<DisassociateFleetResult, DisassociateFleetError>

Disassociates a fleet from a stack.

fn expire_session(
&self,
input: &ExpireSessionRequest
) -> Result<ExpireSessionResult, ExpireSessionError>

This operation immediately stops a streaming session.

fn list_associated_fleets(
&self,
input: &ListAssociatedFleetsRequest
) -> Result<ListAssociatedFleetsResult, ListAssociatedFleetsError>

Lists all fleets associated with the stack.

fn list_associated_stacks(
&self,
input: &ListAssociatedStacksRequest
) -> Result<ListAssociatedStacksResult, ListAssociatedStacksError>

Lists all stacks to which the specified fleet is associated.

fn start_fleet(
&self,
input: &StartFleetRequest
) -> Result<StartFleetResult, StartFleetError>

Starts a fleet.

fn stop_fleet(
&self,
input: &StopFleetRequest
) -> Result<StopFleetResult, StopFleetError>

Stops a fleet.

fn update_fleet(
&self,
input: &UpdateFleetRequest
) -> Result<UpdateFleetResult, UpdateFleetError>

Updates an existing fleet. All the attributes except the fleet name can be updated in the STOPPED state. When a fleet is in the RUNNING state, only DisplayName and ComputeCapacity can be updated. A fleet cannot be updated in a status of STARTING or STOPPING.

fn update_stack(
&self,
input: &UpdateStackRequest
) -> Result<UpdateStackResult, UpdateStackError>

Updates the specified fields in the stack with the specified name.