Crate rusoto_appstream

Amazon AppStream 2.0
API documentation for Amazon AppStream 2.0.

If you're using the service, you're probably looking for AppStreamClient and AppStream.

Structs

AppStreamClient
A client for the Amazon AppStream API.

Application
An entry for a single application in the application catalog.

AssociateFleetRequest

AssociateFleetResult

ComputeCapacity
The capacity configuration for the fleet.

ComputeCapacityStatus
The capacity information for the fleet.

CreateFleetRequest
Contains the parameters for the new fleet to create.

CreateFleetResult

CreateStackRequest

CreateStackResult

CreateStreamingURLRequest

CreateStreamingURLResult

DeleteFleetRequest

DeleteFleetResult

DeleteStackRequest

DeleteStackResult

DescribeFleetsRequest

DescribeFleetsResult

DescribeImagesRequest

DescribeImagesResult

DescribeSessionsRequest

DescribeSessionsResult

DescribeStacksRequest

DescribeStacksResult

DisassociateFleetRequest

DisassociateFleetResult

ExpireSessionRequest

ExpireSessionResult

Fleet
Contains the parameters for a fleet.

FleetError
The details of the fleet error.

Image
New streaming instances are booted from images. The image stores the application catalog and is connected to fleets.

ImageStateChangeReason
The reason why the last state change occurred.

ListAssociatedFleetsRequest

ListAssociatedFleetsResult
The response from a successful operation.

ListAssociatedStacksRequest

ListAssociatedStacksResult
The response from a successful operation.

Session
Contains the parameters for a streaming session.

Stack
Details about a stack.

StackError
Contains the parameters for a stack error.

StartFleetRequest

StartFleetResult

StopFleetRequest

StopFleetResult

StorageConnector
Contains the parameters for a storage connector.

UpdateFleetRequest

UpdateFleetResult

UpdateStackRequest

UpdateStackResult

VpcConfig
VPC configuration information.

Enums

AssociateFleetError
Errors returned by AssociateFleet

CreateFleetError
Errors returned by CreateFleet

CreateStackError
Errors returned by CreateStack

CreateStreamingURLError
Errors returned by CreateStreamingURL

DeleteFleetError
Errors returned by DeleteFleet

DeleteStackError
Errors returned by DeleteStack

DescribeFleetsError
Errors returned by DescribeFleets

DescribeImagesError
Errors returned by DescribeImages

DescribeSessionsError
Errors returned by DescribeSessions

DescribeStacksError
Errors returned by DescribeStacks

DisassociateFleetError
Errors returned by DisassociateFleet

ExpireSessionError
Errors returned by ExpireSession

ListAssociatedFleetsError
Errors returned by ListAssociatedFleets

ListAssociatedStacksError
Errors returned by ListAssociatedStacks

StartFleetError
Errors returned by StartFleet

StopFleetError
Errors returned by StopFleet

UpdateFleetError
Errors returned by UpdateFleet

UpdateStackError
Errors returned by UpdateStack

Traits

AppStream
Trait representing the capabilities of the Amazon AppStream API. Amazon AppStream clients implement this trait.