Crate rusoto_appstream

Amazon AppStream 2.0
You can use Amazon AppStream 2.0 to stream desktop applications to any device running a web browser, without rewriting them.

If you're using the service, you're probably looking for AppStreamClient and AppStream.

Structs

AppStreamClient
A client for the Amazon AppStream API.

Application
Describes an application in the application catalog.

AssociateFleetRequest

AssociateFleetResult

ComputeCapacity
Describes the capacity for a fleet.

ComputeCapacityStatus
Describes the capacity status for a fleet.

CopyImageRequest

CopyImageResponse

CreateDirectoryConfigRequest

CreateDirectoryConfigResult

CreateFleetRequest

CreateFleetResult

CreateImageBuilderRequest

CreateImageBuilderResult

CreateImageBuilderStreamingURLRequest

CreateImageBuilderStreamingURLResult

CreateStackRequest

CreateStackResult

CreateStreamingURLRequest

CreateStreamingURLResult

DeleteDirectoryConfigRequest

DeleteDirectoryConfigResult

DeleteFleetRequest

DeleteFleetResult

DeleteImageBuilderRequest

DeleteImageBuilderResult

DeleteImagePermissionsRequest

DeleteImagePermissionsResult

DeleteImageRequest

DeleteImageResult

DeleteStackRequest

DeleteStackResult

DescribeDirectoryConfigsRequest

DescribeDirectoryConfigsResult

DescribeFleetsRequest

DescribeFleetsResult

DescribeImageBuildersRequest

DescribeImageBuildersResult

DescribeImagePermissionsRequest

DescribeImagePermissionsResult

DescribeImagesRequest

DescribeImagesResult

DescribeSessionsRequest

DescribeSessionsResult

DescribeStacksRequest

DescribeStacksResult

DirectoryConfig
Configuration information for the directory used to join domains.

DisassociateFleetRequest

DisassociateFleetResult

DomainJoinInfo
Contains the information needed to join a Microsoft Active Directory domain.

ExpireSessionRequest

ExpireSessionResult

Fleet
Contains the parameters for a fleet.

FleetError
Describes a fleet error.

Image
Describes an image.

ImageBuilder
Describes a streaming instance used for editing an image. New images are created from a snapshot through an image builder.

ImageBuilderStateChangeReason
Describes the reason why the last image builder state change occurred.

ImagePermissions
Describes the permissions for an image.

ImageStateChangeReason
Describes the reason why the last image state change occurred.

ListAssociatedFleetsRequest

ListAssociatedFleetsResult

ListAssociatedStacksRequest

ListAssociatedStacksResult

ListTagsForResourceRequest

ListTagsForResourceResponse

NetworkAccessConfiguration
The network details of the fleet instance for the streaming session.

ResourceError
Describes a resource error.

ServiceAccountCredentials
Describes the credentials for the service account used by the streaming instance to connect to the directory.

Session
Describes a streaming session.

SharedImagePermissions
Describes the permissions that are available to the specified AWS account for a shared image.

Stack
Describes a stack.

StackError
Describes a stack error.

StartFleetRequest

StartFleetResult

StartImageBuilderRequest

StartImageBuilderResult

StopFleetRequest

StopFleetResult

StopImageBuilderRequest

StopImageBuilderResult

StorageConnector
Describes a connector to enable persistent storage for users.

TagResourceRequest

TagResourceResponse

UntagResourceRequest

UntagResourceResponse

UpdateDirectoryConfigRequest

UpdateDirectoryConfigResult

UpdateFleetRequest

UpdateFleetResult

UpdateImagePermissionsRequest

UpdateImagePermissionsResult

UpdateStackRequest

UpdateStackResult

UserSetting
Describes an action and whether the action is enabled or disabled for users during their streaming sessions.

VpcConfig
Describes VPC configuration information.

Enums

AssociateFleetError
Errors returned by AssociateFleet

CopyImageError
Errors returned by CopyImage

CreateDirectoryConfigError
Errors returned by CreateDirectoryConfig

CreateFleetError
Errors returned by CreateFleet

CreateImageBuilderError
Errors returned by CreateImageBuilder

CreateImageBuilderStreamingURLError
Errors returned by CreateImageBuilderStreamingURL

CreateStackError
Errors returned by CreateStack

CreateStreamingURLError
Errors returned by CreateStreamingURL

DeleteDirectoryConfigError
Errors returned by DeleteDirectoryConfig

DeleteFleetError
Errors returned by DeleteFleet

DeleteImageBuilderError
Errors returned by DeleteImageBuilder

DeleteImageError
Errors returned by DeleteImage

DeleteImagePermissionsError
Errors returned by DeleteImagePermissions

DeleteStackError
Errors returned by DeleteStack

DescribeDirectoryConfigsError
Errors returned by DescribeDirectoryConfigs

DescribeFleetsError
Errors returned by DescribeFleets

DescribeImageBuildersError
Errors returned by DescribeImageBuilders

DescribeImagePermissionsError
Errors returned by DescribeImagePermissions

DescribeImagesError
Errors returned by DescribeImages

DescribeSessionsError
Errors returned by DescribeSessions

DescribeStacksError
Errors returned by DescribeStacks

DisassociateFleetError
Errors returned by DisassociateFleet

ExpireSessionError
Errors returned by ExpireSession

ListAssociatedFleetsError
Errors returned by ListAssociatedFleets

ListAssociatedStacksError
Errors returned by ListAssociatedStacks

ListTagsForResourceError
Errors returned by ListTagsForResource

StartFleetError
Errors returned by StartFleet

StartImageBuilderError
Errors returned by StartImageBuilder

StopFleetError
Errors returned by StopFleet

StopImageBuilderError
Errors returned by StopImageBuilder

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

UpdateDirectoryConfigError
Errors returned by UpdateDirectoryConfig

UpdateFleetError
Errors returned by UpdateFleet

UpdateImagePermissionsError
Errors returned by UpdateImagePermissions

UpdateStackError
Errors returned by UpdateStack

Traits

AppStream
Trait representing the capabilities of the Amazon AppStream API. Amazon AppStream clients implement this trait.