Struct rusoto_appstream::AppStreamClient

pub struct AppStreamClient { /* fields omitted */ }

A client for the Amazon AppStream API.

Methods

impl AppStreamClient

pub fn new(region: Region) -> AppStreamClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> AppStreamClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,