Trait rusoto_appstream::AppStream

pub trait AppStream {
    fn associate_fleet(
        &self, 
        input: AssociateFleetRequest
    ) -> RusotoFuture<AssociateFleetResult, AssociateFleetError>;

    fn batch_associate_user_stack(
        &self, 
        input: BatchAssociateUserStackRequest
    ) -> RusotoFuture<BatchAssociateUserStackResult, BatchAssociateUserStackError>;

    fn batch_disassociate_user_stack(
        &self, 
        input: BatchDisassociateUserStackRequest
    ) -> RusotoFuture<BatchDisassociateUserStackResult, BatchDisassociateUserStackError>;

    fn copy_image(
        &self, 
        input: CopyImageRequest
    ) -> RusotoFuture<CopyImageResponse, CopyImageError>;

    fn create_directory_config(
        &self, 
        input: CreateDirectoryConfigRequest
    ) -> RusotoFuture<CreateDirectoryConfigResult, CreateDirectoryConfigError>;

    fn create_fleet(
        &self, 
        input: CreateFleetRequest
    ) -> RusotoFuture<CreateFleetResult, CreateFleetError>;

    fn create_image_builder(
        &self, 
        input: CreateImageBuilderRequest
    ) -> RusotoFuture<CreateImageBuilderResult, CreateImageBuilderError>;

    fn create_image_builder_streaming_url(
        &self, 
        input: CreateImageBuilderStreamingURLRequest
    ) -> RusotoFuture<CreateImageBuilderStreamingURLResult, CreateImageBuilderStreamingURLError>;

    fn create_stack(
        &self, 
        input: CreateStackRequest
    ) -> RusotoFuture<CreateStackResult, CreateStackError>;

    fn create_streaming_url(
        &self, 
        input: CreateStreamingURLRequest
    ) -> RusotoFuture<CreateStreamingURLResult, CreateStreamingURLError>;

    fn create_usage_report_subscription(
        &self
    ) -> RusotoFuture<CreateUsageReportSubscriptionResult, CreateUsageReportSubscriptionError>;

    fn create_user(
        &self, 
        input: CreateUserRequest
    ) -> RusotoFuture<CreateUserResult, CreateUserError>;

    fn delete_directory_config(
        &self, 
        input: DeleteDirectoryConfigRequest
    ) -> RusotoFuture<DeleteDirectoryConfigResult, DeleteDirectoryConfigError>;

    fn delete_fleet(
        &self, 
        input: DeleteFleetRequest
    ) -> RusotoFuture<DeleteFleetResult, DeleteFleetError>;

    fn delete_image(
        &self, 
        input: DeleteImageRequest
    ) -> RusotoFuture<DeleteImageResult, DeleteImageError>;

    fn delete_image_builder(
        &self, 
        input: DeleteImageBuilderRequest
    ) -> RusotoFuture<DeleteImageBuilderResult, DeleteImageBuilderError>;

    fn delete_image_permissions(
        &self, 
        input: DeleteImagePermissionsRequest
    ) -> RusotoFuture<DeleteImagePermissionsResult, DeleteImagePermissionsError>;

    fn delete_stack(
        &self, 
        input: DeleteStackRequest
    ) -> RusotoFuture<DeleteStackResult, DeleteStackError>;

    fn delete_usage_report_subscription(
        &self
    ) -> RusotoFuture<DeleteUsageReportSubscriptionResult, DeleteUsageReportSubscriptionError>;

    fn delete_user(
        &self, 
        input: DeleteUserRequest
    ) -> RusotoFuture<DeleteUserResult, DeleteUserError>;

    fn describe_directory_configs(
        &self, 
        input: DescribeDirectoryConfigsRequest
    ) -> RusotoFuture<DescribeDirectoryConfigsResult, DescribeDirectoryConfigsError>;

    fn describe_fleets(
        &self, 
        input: DescribeFleetsRequest
    ) -> RusotoFuture<DescribeFleetsResult, DescribeFleetsError>;

    fn describe_image_builders(
        &self, 
        input: DescribeImageBuildersRequest
    ) -> RusotoFuture<DescribeImageBuildersResult, DescribeImageBuildersError>;

    fn describe_image_permissions(
        &self, 
        input: DescribeImagePermissionsRequest
    ) -> RusotoFuture<DescribeImagePermissionsResult, DescribeImagePermissionsError>;

    fn describe_images(
        &self, 
        input: DescribeImagesRequest
    ) -> RusotoFuture<DescribeImagesResult, DescribeImagesError>;

    fn describe_sessions(
        &self, 
        input: DescribeSessionsRequest
    ) -> RusotoFuture<DescribeSessionsResult, DescribeSessionsError>;

    fn describe_stacks(
        &self, 
        input: DescribeStacksRequest
    ) -> RusotoFuture<DescribeStacksResult, DescribeStacksError>;

    fn describe_usage_report_subscriptions(
        &self, 
        input: DescribeUsageReportSubscriptionsRequest
    ) -> RusotoFuture<DescribeUsageReportSubscriptionsResult, DescribeUsageReportSubscriptionsError>;

    fn describe_user_stack_associations(
        &self, 
        input: DescribeUserStackAssociationsRequest
    ) -> RusotoFuture<DescribeUserStackAssociationsResult, DescribeUserStackAssociationsError>;

    fn describe_users(
        &self, 
        input: DescribeUsersRequest
    ) -> RusotoFuture<DescribeUsersResult, DescribeUsersError>;

    fn disable_user(
        &self, 
        input: DisableUserRequest
    ) -> RusotoFuture<DisableUserResult, DisableUserError>;

    fn disassociate_fleet(
        &self, 
        input: DisassociateFleetRequest
    ) -> RusotoFuture<DisassociateFleetResult, DisassociateFleetError>;

    fn enable_user(
        &self, 
        input: EnableUserRequest
    ) -> RusotoFuture<EnableUserResult, EnableUserError>;

    fn expire_session(
        &self, 
        input: ExpireSessionRequest
    ) -> RusotoFuture<ExpireSessionResult, ExpireSessionError>;

    fn list_associated_fleets(
        &self, 
        input: ListAssociatedFleetsRequest
    ) -> RusotoFuture<ListAssociatedFleetsResult, ListAssociatedFleetsError>;

    fn list_associated_stacks(
        &self, 
        input: ListAssociatedStacksRequest
    ) -> RusotoFuture<ListAssociatedStacksResult, ListAssociatedStacksError>;

    fn list_tags_for_resource(
        &self, 
        input: ListTagsForResourceRequest
    ) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>;

    fn start_fleet(
        &self, 
        input: StartFleetRequest
    ) -> RusotoFuture<StartFleetResult, StartFleetError>;

    fn start_image_builder(
        &self, 
        input: StartImageBuilderRequest
    ) -> RusotoFuture<StartImageBuilderResult, StartImageBuilderError>;

    fn stop_fleet(
        &self, 
        input: StopFleetRequest
    ) -> RusotoFuture<StopFleetResult, StopFleetError>;

    fn stop_image_builder(
        &self, 
        input: StopImageBuilderRequest
    ) -> RusotoFuture<StopImageBuilderResult, StopImageBuilderError>;

    fn tag_resource(
        &self, 
        input: TagResourceRequest
    ) -> RusotoFuture<TagResourceResponse, TagResourceError>;

    fn untag_resource(
        &self, 
        input: UntagResourceRequest
    ) -> RusotoFuture<UntagResourceResponse, UntagResourceError>;

    fn update_directory_config(
        &self, 
        input: UpdateDirectoryConfigRequest
    ) -> RusotoFuture<UpdateDirectoryConfigResult, UpdateDirectoryConfigError>;

    fn update_fleet(
        &self, 
        input: UpdateFleetRequest
    ) -> RusotoFuture<UpdateFleetResult, UpdateFleetError>;

    fn update_image_permissions(
        &self, 
        input: UpdateImagePermissionsRequest
    ) -> RusotoFuture<UpdateImagePermissionsResult, UpdateImagePermissionsError>;

    fn update_stack(
        &self, 
        input: UpdateStackRequest
    ) -> RusotoFuture<UpdateStackResult, UpdateStackError>;
}

Trait representing the capabilities of the Amazon AppStream API. Amazon AppStream clients implement this trait.

Required methods

fn associate_fleet(
&self,
input: AssociateFleetRequest
) -> RusotoFuture<AssociateFleetResult, AssociateFleetError>

Associates the specified fleet with the specified stack.

fn batch_associate_user_stack(
&self,
input: BatchAssociateUserStackRequest
) -> RusotoFuture<BatchAssociateUserStackResult, BatchAssociateUserStackError>

Associates the specified users with the specified stacks. Users in a user pool cannot be assigned to stacks with fleets that are joined to an Active Directory domain.

fn batch_disassociate_user_stack(
&self,
input: BatchDisassociateUserStackRequest
) -> RusotoFuture<BatchDisassociateUserStackResult, BatchDisassociateUserStackError>

Disassociates the specified users from the specified stacks.

fn copy_image(
&self,
input: CopyImageRequest
) -> RusotoFuture<CopyImageResponse, CopyImageError>

Copies the image within the same region or to a new region within the same AWS account. Note that any tags you added to the image will not be copied.

fn create_directory_config(
&self,
input: CreateDirectoryConfigRequest
) -> RusotoFuture<CreateDirectoryConfigResult, CreateDirectoryConfigError>

Creates a Directory Config object in AppStream 2.0. This object includes the configuration information required to join fleets and image builders to Microsoft Active Directory domains.

fn create_fleet(
&self,
input: CreateFleetRequest
) -> RusotoFuture<CreateFleetResult, CreateFleetError>

Creates a fleet. A fleet consists of streaming instances that run a specified image.

fn create_image_builder(
&self,
input: CreateImageBuilderRequest
) -> RusotoFuture<CreateImageBuilderResult, CreateImageBuilderError>

Creates an image builder. An image builder is a virtual machine that is used to create an image.

The initial state of the builder is PENDING. When it is ready, the state is RUNNING.

fn create_image_builder_streaming_url(
&self,
input: CreateImageBuilderStreamingURLRequest
) -> RusotoFuture<CreateImageBuilderStreamingURLResult, CreateImageBuilderStreamingURLError>

Creates a URL to start an image builder streaming session.

fn create_stack(
&self,
input: CreateStackRequest
) -> RusotoFuture<CreateStackResult, CreateStackError>

Creates a stack to start streaming applications to users. A stack consists of an associated fleet, user access policies, and storage configurations.

fn create_streaming_url(
&self,
input: CreateStreamingURLRequest
) -> RusotoFuture<CreateStreamingURLResult, CreateStreamingURLError>

Creates a temporary URL to start an AppStream 2.0 streaming session for the specified user. A streaming URL enables application streaming to be tested without user setup.

fn create_usage_report_subscription(
&self
) -> RusotoFuture<CreateUsageReportSubscriptionResult, CreateUsageReportSubscriptionError>

Creates a usage report subscription. Usage reports are generated daily.

fn create_user(
&self,
input: CreateUserRequest
) -> RusotoFuture<CreateUserResult, CreateUserError>

Creates a new user in the user pool.

fn delete_directory_config(
&self,
input: DeleteDirectoryConfigRequest
) -> RusotoFuture<DeleteDirectoryConfigResult, DeleteDirectoryConfigError>

Deletes the specified Directory Config object from AppStream 2.0. This object includes the information required to join streaming instances to an Active Directory domain.

fn delete_fleet(
&self,
input: DeleteFleetRequest
) -> RusotoFuture<DeleteFleetResult, DeleteFleetError>

Deletes the specified fleet.

fn delete_image(
&self,
input: DeleteImageRequest
) -> RusotoFuture<DeleteImageResult, DeleteImageError>

Deletes the specified image. You cannot delete an image when it is in use. After you delete an image, you cannot provision new capacity using the image.

fn delete_image_builder(
&self,
input: DeleteImageBuilderRequest
) -> RusotoFuture<DeleteImageBuilderResult, DeleteImageBuilderError>

Deletes the specified image builder and releases the capacity.

fn delete_image_permissions(
&self,
input: DeleteImagePermissionsRequest
) -> RusotoFuture<DeleteImagePermissionsResult, DeleteImagePermissionsError>

Deletes permissions for the specified private image. After you delete permissions for an image, AWS accounts to which you previously granted these permissions can no longer use the image.

fn delete_stack(
&self,
input: DeleteStackRequest
) -> RusotoFuture<DeleteStackResult, DeleteStackError>

Deletes the specified stack. After the stack is deleted, the application streaming environment provided by the stack is no longer available to users. Also, any reservations made for application streaming sessions for the stack are released.

fn delete_usage_report_subscription(
&self
) -> RusotoFuture<DeleteUsageReportSubscriptionResult, DeleteUsageReportSubscriptionError>

Disables usage report generation.

fn delete_user(
&self,
input: DeleteUserRequest
) -> RusotoFuture<DeleteUserResult, DeleteUserError>

Deletes a user from the user pool.

fn describe_directory_configs(
&self,
input: DescribeDirectoryConfigsRequest
) -> RusotoFuture<DescribeDirectoryConfigsResult, DescribeDirectoryConfigsError>

Retrieves a list that describes one or more specified Directory Config objects for AppStream 2.0, if the names for these objects are provided. Otherwise, all Directory Config objects in the account are described. These objects include the configuration information required to join fleets and image builders to Microsoft Active Directory domains.

Although the response syntax in this topic includes the account password, this password is not returned in the actual response.

fn describe_fleets(
&self,
input: DescribeFleetsRequest
) -> RusotoFuture<DescribeFleetsResult, DescribeFleetsError>

Retrieves a list that describes one or more specified fleets, if the fleet names are provided. Otherwise, all fleets in the account are described.

fn describe_image_builders(
&self,
input: DescribeImageBuildersRequest
) -> RusotoFuture<DescribeImageBuildersResult, DescribeImageBuildersError>

Retrieves a list that describes one or more specified image builders, if the image builder names are provided. Otherwise, all image builders in the account are described.

fn describe_image_permissions(
&self,
input: DescribeImagePermissionsRequest
) -> RusotoFuture<DescribeImagePermissionsResult, DescribeImagePermissionsError>

Retrieves a list that describes the permissions for shared AWS account IDs on a private image that you own.

fn describe_images(
&self,
input: DescribeImagesRequest
) -> RusotoFuture<DescribeImagesResult, DescribeImagesError>

Retrieves a list that describes one or more specified images, if the image names or image ARNs are provided. Otherwise, all images in the account are described.

fn describe_sessions(
&self,
input: DescribeSessionsRequest
) -> RusotoFuture<DescribeSessionsResult, DescribeSessionsError>

Retrieves a list that describes the streaming sessions for a specified stack and fleet. If a UserId is provided for the stack and fleet, only streaming sessions for that user are described. If an authentication type is not provided, the default is to authenticate users using a streaming URL.

fn describe_stacks(
&self,
input: DescribeStacksRequest
) -> RusotoFuture<DescribeStacksResult, DescribeStacksError>

Retrieves a list that describes one or more specified stacks, if the stack names are provided. Otherwise, all stacks in the account are described.

fn describe_usage_report_subscriptions(
&self,
input: DescribeUsageReportSubscriptionsRequest
) -> RusotoFuture<DescribeUsageReportSubscriptionsResult, DescribeUsageReportSubscriptionsError>

Retrieves a list that describes one or more usage report subscriptions.

fn describe_user_stack_associations(
&self,
input: DescribeUserStackAssociationsRequest
) -> RusotoFuture<DescribeUserStackAssociationsResult, DescribeUserStackAssociationsError>

Retrieves a list that describes the UserStackAssociation objects. You must specify either or both of the following:

The stack name

The user name (email address of the user associated with the stack) and the authentication type for the user

fn describe_users(
&self,
input: DescribeUsersRequest
) -> RusotoFuture<DescribeUsersResult, DescribeUsersError>

Retrieves a list that describes one or more specified users in the user pool.

fn disable_user(
&self,
input: DisableUserRequest
) -> RusotoFuture<DisableUserResult, DisableUserError>

Disables the specified user in the user pool. Users can't sign in to AppStream 2.0 until they are re-enabled. This action does not delete the user.

fn disassociate_fleet(
&self,
input: DisassociateFleetRequest
) -> RusotoFuture<DisassociateFleetResult, DisassociateFleetError>

Disassociates the specified fleet from the specified stack.

fn enable_user(
&self,
input: EnableUserRequest
) -> RusotoFuture<EnableUserResult, EnableUserError>

Enables a user in the user pool. After being enabled, users can sign in to AppStream 2.0 and open applications from the stacks to which they are assigned.

fn expire_session(
&self,
input: ExpireSessionRequest
) -> RusotoFuture<ExpireSessionResult, ExpireSessionError>

Immediately stops the specified streaming session.

fn list_associated_fleets(
&self,
input: ListAssociatedFleetsRequest
) -> RusotoFuture<ListAssociatedFleetsResult, ListAssociatedFleetsError>

Retrieves the name of the fleet that is associated with the specified stack.

fn list_associated_stacks(
&self,
input: ListAssociatedStacksRequest
) -> RusotoFuture<ListAssociatedStacksResult, ListAssociatedStacksError>

Retrieves the name of the stack with which the specified fleet is associated.

fn list_tags_for_resource(
&self,
input: ListTagsForResourceRequest
) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>

Retrieves a list of all tags for the specified AppStream 2.0 resource. You can tag AppStream 2.0 image builders, images, fleets, and stacks.

For more information about tags, see Tagging Your Resources in the Amazon AppStream 2.0 Developer Guide.

fn start_fleet(
&self,
input: StartFleetRequest
) -> RusotoFuture<StartFleetResult, StartFleetError>

Starts the specified fleet.

fn start_image_builder(
&self,
input: StartImageBuilderRequest
) -> RusotoFuture<StartImageBuilderResult, StartImageBuilderError>

Starts the specified image builder.

fn stop_fleet(
&self,
input: StopFleetRequest
) -> RusotoFuture<StopFleetResult, StopFleetError>

Stops the specified fleet.

fn stop_image_builder(
&self,
input: StopImageBuilderRequest
) -> RusotoFuture<StopImageBuilderResult, StopImageBuilderError>

Stops the specified image builder.

fn tag_resource(
&self,
input: TagResourceRequest
) -> RusotoFuture<TagResourceResponse, TagResourceError>

Adds or overwrites one or more tags for the specified AppStream 2.0 resource. You can tag AppStream 2.0 image builders, images, fleets, and stacks.

Each tag consists of a key and an optional value. If a resource already has a tag with the same key, this operation updates its value.

To list the current tags for your resources, use ListTagsForResource. To disassociate tags from your resources, use UntagResource.

For more information about tags, see Tagging Your Resources in the Amazon AppStream 2.0 Developer Guide.

fn untag_resource(
&self,
input: UntagResourceRequest
) -> RusotoFuture<UntagResourceResponse, UntagResourceError>

Disassociates one or more specified tags from the specified AppStream 2.0 resource.

To list the current tags for your resources, use ListTagsForResource.

For more information about tags, see Tagging Your Resources in the Amazon AppStream 2.0 Developer Guide.

fn update_directory_config(
&self,
input: UpdateDirectoryConfigRequest
) -> RusotoFuture<UpdateDirectoryConfigResult, UpdateDirectoryConfigError>

Updates the specified Directory Config object in AppStream 2.0. This object includes the configuration information required to join fleets and image builders to Microsoft Active Directory domains.

fn update_fleet(
&self,
input: UpdateFleetRequest
) -> RusotoFuture<UpdateFleetResult, UpdateFleetError>

Updates the specified fleet.

If the fleet is in the STOPPED state, you can update any attribute except the fleet name. If the fleet is in the RUNNING state, you can update the DisplayName, ComputeCapacity, ImageARN, ImageName, and DisconnectTimeoutInSeconds attributes. If the fleet is in the STARTING or STOPPING state, you can't update it.

fn update_image_permissions(
&self,
input: UpdateImagePermissionsRequest
) -> RusotoFuture<UpdateImagePermissionsResult, UpdateImagePermissionsError>

Adds or updates permissions for the specified private image.

fn update_stack(
&self,
input: UpdateStackRequest
) -> RusotoFuture<UpdateStackResult, UpdateStackError>

Updates the specified fields for the specified stack.

Loading content...