Trait rusoto_appstream::AppStream

pub trait AppStream {
    fn associate_fleet(
        &self, 
        input: AssociateFleetRequest
    ) -> RusotoFuture<AssociateFleetResult, AssociateFleetError>;

    fn copy_image(
        &self, 
        input: CopyImageRequest
    ) -> RusotoFuture<CopyImageResponse, CopyImageError>;

    fn create_directory_config(
        &self, 
        input: CreateDirectoryConfigRequest
    ) -> RusotoFuture<CreateDirectoryConfigResult, CreateDirectoryConfigError>;

    fn create_fleet(
        &self, 
        input: CreateFleetRequest
    ) -> RusotoFuture<CreateFleetResult, CreateFleetError>;

    fn create_image_builder(
        &self, 
        input: CreateImageBuilderRequest
    ) -> RusotoFuture<CreateImageBuilderResult, CreateImageBuilderError>;

    fn create_image_builder_streaming_url(
        &self, 
        input: CreateImageBuilderStreamingURLRequest
    ) -> RusotoFuture<CreateImageBuilderStreamingURLResult, CreateImageBuilderStreamingURLError>;

    fn create_stack(
        &self, 
        input: CreateStackRequest
    ) -> RusotoFuture<CreateStackResult, CreateStackError>;

    fn create_streaming_url(
        &self, 
        input: CreateStreamingURLRequest
    ) -> RusotoFuture<CreateStreamingURLResult, CreateStreamingURLError>;

    fn delete_directory_config(
        &self, 
        input: DeleteDirectoryConfigRequest
    ) -> RusotoFuture<DeleteDirectoryConfigResult, DeleteDirectoryConfigError>;

    fn delete_fleet(
        &self, 
        input: DeleteFleetRequest
    ) -> RusotoFuture<DeleteFleetResult, DeleteFleetError>;

    fn delete_image(
        &self, 
        input: DeleteImageRequest
    ) -> RusotoFuture<DeleteImageResult, DeleteImageError>;

    fn delete_image_builder(
        &self, 
        input: DeleteImageBuilderRequest
    ) -> RusotoFuture<DeleteImageBuilderResult, DeleteImageBuilderError>;

    fn delete_stack(
        &self, 
        input: DeleteStackRequest
    ) -> RusotoFuture<DeleteStackResult, DeleteStackError>;

    fn describe_directory_configs(
        &self, 
        input: DescribeDirectoryConfigsRequest
    ) -> RusotoFuture<DescribeDirectoryConfigsResult, DescribeDirectoryConfigsError>;

    fn describe_fleets(
        &self, 
        input: DescribeFleetsRequest
    ) -> RusotoFuture<DescribeFleetsResult, DescribeFleetsError>;

    fn describe_image_builders(
        &self, 
        input: DescribeImageBuildersRequest
    ) -> RusotoFuture<DescribeImageBuildersResult, DescribeImageBuildersError>;

    fn describe_images(
        &self, 
        input: DescribeImagesRequest
    ) -> RusotoFuture<DescribeImagesResult, DescribeImagesError>;

    fn describe_sessions(
        &self, 
        input: DescribeSessionsRequest
    ) -> RusotoFuture<DescribeSessionsResult, DescribeSessionsError>;

    fn describe_stacks(
        &self, 
        input: DescribeStacksRequest
    ) -> RusotoFuture<DescribeStacksResult, DescribeStacksError>;

    fn disassociate_fleet(
        &self, 
        input: DisassociateFleetRequest
    ) -> RusotoFuture<DisassociateFleetResult, DisassociateFleetError>;

    fn expire_session(
        &self, 
        input: ExpireSessionRequest
    ) -> RusotoFuture<ExpireSessionResult, ExpireSessionError>;

    fn list_associated_fleets(
        &self, 
        input: ListAssociatedFleetsRequest
    ) -> RusotoFuture<ListAssociatedFleetsResult, ListAssociatedFleetsError>;

    fn list_associated_stacks(
        &self, 
        input: ListAssociatedStacksRequest
    ) -> RusotoFuture<ListAssociatedStacksResult, ListAssociatedStacksError>;

    fn list_tags_for_resource(
        &self, 
        input: ListTagsForResourceRequest
    ) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>;

    fn start_fleet(
        &self, 
        input: StartFleetRequest
    ) -> RusotoFuture<StartFleetResult, StartFleetError>;

    fn start_image_builder(
        &self, 
        input: StartImageBuilderRequest
    ) -> RusotoFuture<StartImageBuilderResult, StartImageBuilderError>;

    fn stop_fleet(
        &self, 
        input: StopFleetRequest
    ) -> RusotoFuture<StopFleetResult, StopFleetError>;

    fn stop_image_builder(
        &self, 
        input: StopImageBuilderRequest
    ) -> RusotoFuture<StopImageBuilderResult, StopImageBuilderError>;

    fn tag_resource(
        &self, 
        input: TagResourceRequest
    ) -> RusotoFuture<TagResourceResponse, TagResourceError>;

    fn untag_resource(
        &self, 
        input: UntagResourceRequest
    ) -> RusotoFuture<UntagResourceResponse, UntagResourceError>;

    fn update_directory_config(
        &self, 
        input: UpdateDirectoryConfigRequest
    ) -> RusotoFuture<UpdateDirectoryConfigResult, UpdateDirectoryConfigError>;

    fn update_fleet(
        &self, 
        input: UpdateFleetRequest
    ) -> RusotoFuture<UpdateFleetResult, UpdateFleetError>;

    fn update_stack(
        &self, 
        input: UpdateStackRequest
    ) -> RusotoFuture<UpdateStackResult, UpdateStackError>;
}

Trait representing the capabilities of the Amazon AppStream API. Amazon AppStream clients implement this trait.

Required Methods

fn associate_fleet(
&self,
input: AssociateFleetRequest
) -> RusotoFuture<AssociateFleetResult, AssociateFleetError>

Associates the specified fleet with the specified stack.

fn copy_image(
&self,
input: CopyImageRequest
) -> RusotoFuture<CopyImageResponse, CopyImageError>

Copies the image within the same region or to a new region within the same AWS account. Note that any tags you added to the image will not be copied.

fn create_directory_config(
&self,
input: CreateDirectoryConfigRequest
) -> RusotoFuture<CreateDirectoryConfigResult, CreateDirectoryConfigError>

Creates a directory configuration.

fn create_fleet(
&self,
input: CreateFleetRequest
) -> RusotoFuture<CreateFleetResult, CreateFleetError>

Creates a fleet.

fn create_image_builder(
&self,
input: CreateImageBuilderRequest
) -> RusotoFuture<CreateImageBuilderResult, CreateImageBuilderError>

Creates an image builder.

The initial state of the builder is PENDING. When it is ready, the state is RUNNING.

fn create_image_builder_streaming_url(
&self,
input: CreateImageBuilderStreamingURLRequest
) -> RusotoFuture<CreateImageBuilderStreamingURLResult, CreateImageBuilderStreamingURLError>

Creates a URL to start an image builder streaming session.

fn create_stack(
&self,
input: CreateStackRequest
) -> RusotoFuture<CreateStackResult, CreateStackError>

Creates a stack.

fn create_streaming_url(
&self,
input: CreateStreamingURLRequest
) -> RusotoFuture<CreateStreamingURLResult, CreateStreamingURLError>

Creates a URL to start a streaming session for the specified user.

fn delete_directory_config(
&self,
input: DeleteDirectoryConfigRequest
) -> RusotoFuture<DeleteDirectoryConfigResult, DeleteDirectoryConfigError>

Deletes the specified directory configuration.

fn delete_fleet(
&self,
input: DeleteFleetRequest
) -> RusotoFuture<DeleteFleetResult, DeleteFleetError>

Deletes the specified fleet.

fn delete_image(
&self,
input: DeleteImageRequest
) -> RusotoFuture<DeleteImageResult, DeleteImageError>

Deletes the specified image. You cannot delete an image that is currently in use. After you delete an image, you cannot provision new capacity using the image.

fn delete_image_builder(
&self,
input: DeleteImageBuilderRequest
) -> RusotoFuture<DeleteImageBuilderResult, DeleteImageBuilderError>

Deletes the specified image builder and releases the capacity.

fn delete_stack(
&self,
input: DeleteStackRequest
) -> RusotoFuture<DeleteStackResult, DeleteStackError>

Deletes the specified stack. After this operation completes, the environment can no longer be activated and any reservations made for the stack are released.

fn describe_directory_configs(
&self,
input: DescribeDirectoryConfigsRequest
) -> RusotoFuture<DescribeDirectoryConfigsResult, DescribeDirectoryConfigsError>

Describes the specified directory configurations. Note that although the response syntax in this topic includes the account password, this password is not returned in the actual response.

fn describe_fleets(
&self,
input: DescribeFleetsRequest
) -> RusotoFuture<DescribeFleetsResult, DescribeFleetsError>

Describes the specified fleets or all fleets in the account.

fn describe_image_builders(
&self,
input: DescribeImageBuildersRequest
) -> RusotoFuture<DescribeImageBuildersResult, DescribeImageBuildersError>

Describes the specified image builders or all image builders in the account.

fn describe_images(
&self,
input: DescribeImagesRequest
) -> RusotoFuture<DescribeImagesResult, DescribeImagesError>

Describes the specified images or all images in the account.

fn describe_sessions(
&self,
input: DescribeSessionsRequest
) -> RusotoFuture<DescribeSessionsResult, DescribeSessionsError>

Describes the streaming sessions for the specified stack and fleet. If a user ID is provided, only the streaming sessions for only that user are returned. If an authentication type is not provided, the default is to authenticate users using a streaming URL.

fn describe_stacks(
&self,
input: DescribeStacksRequest
) -> RusotoFuture<DescribeStacksResult, DescribeStacksError>

Describes the specified stacks or all stacks in the account.

fn disassociate_fleet(
&self,
input: DisassociateFleetRequest
) -> RusotoFuture<DisassociateFleetResult, DisassociateFleetError>

Disassociates the specified fleet from the specified stack.

fn expire_session(
&self,
input: ExpireSessionRequest
) -> RusotoFuture<ExpireSessionResult, ExpireSessionError>

Stops the specified streaming session.

fn list_associated_fleets(
&self,
input: ListAssociatedFleetsRequest
) -> RusotoFuture<ListAssociatedFleetsResult, ListAssociatedFleetsError>

Lists the fleets associated with the specified stack.

fn list_associated_stacks(
&self,
input: ListAssociatedStacksRequest
) -> RusotoFuture<ListAssociatedStacksResult, ListAssociatedStacksError>

Lists the stacks associated with the specified fleet.

fn list_tags_for_resource(
&self,
input: ListTagsForResourceRequest
) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>

Lists the tags for the specified AppStream 2.0 resource. You can tag AppStream 2.0 image builders, images, fleets, and stacks.

For more information about tags, see Tagging Your Resources in the Amazon AppStream 2.0 Developer Guide.

fn start_fleet(
&self,
input: StartFleetRequest
) -> RusotoFuture<StartFleetResult, StartFleetError>

Starts the specified fleet.

fn start_image_builder(
&self,
input: StartImageBuilderRequest
) -> RusotoFuture<StartImageBuilderResult, StartImageBuilderError>

Starts the specified image builder.

fn stop_fleet(
&self,
input: StopFleetRequest
) -> RusotoFuture<StopFleetResult, StopFleetError>

Stops the specified fleet.

fn stop_image_builder(
&self,
input: StopImageBuilderRequest
) -> RusotoFuture<StopImageBuilderResult, StopImageBuilderError>

Stops the specified image builder.

fn tag_resource(
&self,
input: TagResourceRequest
) -> RusotoFuture<TagResourceResponse, TagResourceError>

Adds or overwrites one or more tags for the specified AppStream 2.0 resource. You can tag AppStream 2.0 image builders, images, fleets, and stacks.

Each tag consists of a key and an optional value. If a resource already has a tag with the same key, this operation updates its value.

To list the current tags for your resources, use ListTagsForResource. To disassociate tags from your resources, use UntagResource.

For more information about tags, see Tagging Your Resources in the Amazon AppStream 2.0 Developer Guide.

fn untag_resource(
&self,
input: UntagResourceRequest
) -> RusotoFuture<UntagResourceResponse, UntagResourceError>

Disassociates the specified tags from the specified AppStream 2.0 resource.

To list the current tags for your resources, use ListTagsForResource.

For more information about tags, see Tagging Your Resources in the Amazon AppStream 2.0 Developer Guide.

fn update_directory_config(
&self,
input: UpdateDirectoryConfigRequest
) -> RusotoFuture<UpdateDirectoryConfigResult, UpdateDirectoryConfigError>

Updates the specified directory configuration.

fn update_fleet(
&self,
input: UpdateFleetRequest
) -> RusotoFuture<UpdateFleetResult, UpdateFleetError>

Updates the specified fleet.

If the fleet is in the STOPPED state, you can update any attribute except the fleet name. If the fleet is in the RUNNING state, you can update the DisplayName and ComputeCapacity attributes. If the fleet is in the STARTING or STOPPING state, you can't update it.

fn update_stack(
&self,
input: UpdateStackRequest
) -> RusotoFuture<UpdateStackResult, UpdateStackError>

Updates the specified stack.